added optoinal installation

+++ src/machineconfig/setup_windows/repos.ps1
@@ -1,4 +1,7 @@
 
+# Editable installation of crocodile and machineconfig
+# no assumptions on which ve is activate is used for installation, this is dictated by activate_ve.
+
 cd ~
 mkdir code -ErrorAction SilentlyContinue
 cd ~/code
@@ -6,17 +9,16 @@ git clone https://github.com/thisismygitrepo/crocodile.git --depth 4
 git clone https://github.com/thisismygitrepo/machineconfig --depth 4  # Choose browser-based authentication.
 
 
-# CAUTION: The below is preferred over: ~/scripts/activate_ve.ps1 because this script explictly places the repos
-# in the locations as above, and are used as below subsequently
-# $machineconfig_path = (python -c "print(__import__('machineconfig').__file__[:-12])")
-# . $machineconfig_path/scripts/windows/activate_ve.ps1
-# no assumptions on which ve is activate is used for installation, this is dictated by activate_ve.
-# this only solves the matter of where to find activate_ve script
-
 cd ~/code/crocodile
-pip install -e .  # local installation of crocodile (allows for development)
-# pip install -r requirements.txt  # not installed automatically by corocdile.
+
+if (-not (Test-Path variable:CROCODILE_EXTRA)) {
+    Write-Host "⚠️ Using default CROCODILE_EXTRA"
+    pip install -e .
+} else { 
+    Write-Host "➡️ CROCODILE_EXTRA = $CROCODILE_EXTRA"
+    pip install -e .[$CROCODILE_EXTRA]
+}
+
 cd ~/code/machineconfig
 pip install -e .
 echo "Finished setting up repos"
-
